placeholder/requirements.txt
Django==1.8.2
Pillow==2.9.0
MapProxy==1.8.0
